third_party/WebKit/LayoutTests/fast/dom/Window/window-postmessage-clone-deep-array-expected.txt

CONSOLE ERROR: line 11: Uncaught RangeError: Maximum call stack size exceeded
Tests that we support cloning deep(ish) arrays.

PASS: eventData is done of type string
